hf-space/requirements.txt
gradio
fastapi
uvicorn
transformers
torch
Pillow
